services/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
redis==5.0.4
boto3==1.34.84
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
celery[redis]==5.3.6
